backend/requirements.txt
fastapi==0.78.0
uvicorn==0.15.0
python-dotenv==0.15.0
torch==1.11.0
torchvision==0.12